test/python/pyproject.toml
[project]
name = "uid2-docs-python-tests"
version = "0.1.0"
description = "Python tests for UID2 documentation code samples"
requires-python = ">=3.8"
dependencies = [
    "uid2-client>=3.0.0a1",
    "pytest>=7.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.0.0", 
]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--disable-warnings",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]